Move rot_synch class to _test/utils.py
