+++ _test/test_homqcqp.py
@@ -3,21 +3,13 @@ import unittest
 
 import numpy as np
 from poly_matrix import PolyMatrix
+from utils import get_chain_rot_prob, get_loop_rot_prob
 
 from cert_tools import HomQCQP
 from cert_tools.hom_qcqp import greedy_cover
 from cert_tools.linalg_tools import svec
-from cert_tools.problems.rot_synch import RotSynchLoopProblem
 from cert_tools.sdp_solvers import solve_sdp_homqcqp
 from cert_tools.sparse_solvers import solve_clarabel, solve_dsdp
-
-
-def get_chain_rot_prob(N=10, locked_pose=0):
-    return RotSynchLoopProblem(N=N, loop_pose=-1, locked_pose=locked_pose)
-
-
-def get_loop_rot_prob():
-    return RotSynchLoopProblem()
 
 
 class TestHomQCQP(unittest.TestCase):
